move the t3 channel test to separate it class
